敏捷软件开发：原则、模式与实践第19章代码

+++ 软件设计/敏捷开发/敏捷软件开发：原则、模式与实践/Payroll/PayrollCode/cppunit1.5/CppUnit/test/framework/TestResult.h
@@ -5,17 +5,15 @@
 #include <vector>
 
 #ifndef CPPUNIT_GUARDS_H
-#include "Guards.h"
+    #include "Guards.h"
 #endif
 
 #ifndef CPPUNIT_TESTFAILURE_H
-#include "TestFailure.h"
+    #include "TestFailure.h"
 #endif
-
 
 class CppUnitException;
 class Test;
-
 
 /*
  * A TestResult collects the results of executing a test case. It is an 
@@ -39,137 +37,170 @@ class Test;
 
 class TestResult
 {
-    REFERENCEOBJECT (TestResult)
+    REFERENCEOBJECT(TestResult)
 
 public:
-                                        TestResult  ();
-    virtual                             ~TestResult ();
+    TestResult();
+    virtual ~TestResult();
 
-    virtual void                        addError       (Test *test, CppUnitException *e);
-    virtual void                        addFailure     (Test *test, CppUnitException *e);
-    virtual void                        startTest      (Test *test);
-    virtual void                        endTest        (Test *test);
-    virtual int                         runTests       ();
-    virtual int                         testErrors     ();
-    virtual int                         testFailures   ();
-    virtual bool                        wasSuccessful  ();
-    virtual bool                        shouldStop     ();
-    virtual void                        stop           ();
+    virtual void addError(Test* test, CppUnitException* e);
+    virtual void addFailure(Test* test, CppUnitException* e);
+    virtual void startTest(Test* test);
+    virtual void endTest(Test* test);
+    virtual int runTests();
+    virtual int testErrors();
+    virtual int testFailures();
+    virtual bool wasSuccessful();
+    virtual bool shouldStop();
+    virtual void stop();
 
-    virtual std::vector<TestFailure *>& errors         ();
-    virtual std::vector<TestFailure *>& failures       ();
-
+    virtual std::vector<TestFailure*>& errors();
+    virtual std::vector<TestFailure*>& failures();
 
     class SynchronizationObject
     {
     public:
-                                SynchronizationObject  () {}
-        virtual                 ~SynchronizationObject () {}
+        SynchronizationObject()
+        {
+        }
+        virtual ~SynchronizationObject()
+        {
+        }
 
-        virtual void            lock                   () {}
-        virtual void            unlock                 () {}
+        virtual void lock()
+        {
+        }
+        virtual void unlock()
+        {
+        }
     };
 
     class ExclusiveZone
     {
-        SynchronizationObject   *m_syncObject;
+        SynchronizationObject* m_syncObject;
 
     public:
-                                ExclusiveZone (SynchronizationObject *syncObject) 
-                                : m_syncObject (syncObject) 
-                                { m_syncObject->lock (); }
+        ExclusiveZone(SynchronizationObject* syncObject)
+            : m_syncObject(syncObject)
+        {
+            m_syncObject->lock();
+        }
 
-                                ~ExclusiveZone () 
-                                { m_syncObject->unlock (); }
+        ~ExclusiveZone()
+        {
+            m_syncObject->unlock();
+        }
     };
 
 protected:
-    virtual void                setSynchronizationObject (SynchronizationObject *syncObject);
+    virtual void setSynchronizationObject(SynchronizationObject* syncObject);
 
-    std::vector<TestFailure *>  m_errors;
-    std::vector<TestFailure *>  m_failures;
-    int                         m_runTests;
-    bool                        m_stop;
-    SynchronizationObject       *m_syncObject;
-
+    std::vector<TestFailure*> m_errors;
+    std::vector<TestFailure*> m_failures;
+    int m_runTests;
+    bool m_stop;
+    SynchronizationObject* m_syncObject;
 };
 
-
-
 // Construct a TestResult
-inline TestResult::TestResult ()
-: m_syncObject (new SynchronizationObject ())
-{ m_runTests = 0; m_stop = false; }
-
+inline TestResult::TestResult()
+    : m_syncObject(new SynchronizationObject())
+{
+    m_runTests = 0;
+    m_stop = false;
+}
 
 // Adds an error to the list of errors. The passed in exception
 // caused the error
-inline void TestResult::addError (Test *test, CppUnitException *e)
-{ ExclusiveZone zone (m_syncObject); m_errors.push_back (new TestFailure (test, e)); }
-
+inline void TestResult::addError(Test* test, CppUnitException* e)
+{
+    ExclusiveZone zone(m_syncObject);
+    m_errors.push_back(new TestFailure(test, e));
+}
 
 // Adds a failure to the list of failures. The passed in exception
 // caused the failure.
-inline void TestResult::addFailure (Test *test, CppUnitException *e)
-{ ExclusiveZone zone (m_syncObject); m_failures.push_back (new TestFailure (test, e)); }
-
+inline void TestResult::addFailure(Test* test, CppUnitException* e)
+{
+    ExclusiveZone zone(m_syncObject);
+    m_failures.push_back(new TestFailure(test, e));
+}
 
 // Informs the result that a test will be started.
-inline void TestResult::startTest (Test *test)
-{ ExclusiveZone zone (m_syncObject); m_runTests++; }
+inline void TestResult::startTest(Test* test)
+{
+    ExclusiveZone zone(m_syncObject);
+    m_runTests++;
+}
 
-  
 // Informs the result that a test was completed.
-inline void TestResult::endTest (Test *test)
-{ ExclusiveZone zone (m_syncObject); }
-
+inline void TestResult::endTest(Test* test)
+{
+    ExclusiveZone zone(m_syncObject);
+}
 
 // Gets the number of run tests.
-inline int TestResult::runTests ()
-{ ExclusiveZone zone (m_syncObject); return m_runTests; }
-
+inline int TestResult::runTests()
+{
+    ExclusiveZone zone(m_syncObject);
+    return m_runTests;
+}
 
 // Gets the number of detected errors.
-inline int TestResult::testErrors ()
-{ ExclusiveZone zone (m_syncObject); return m_errors.size (); }
-
+inline int TestResult::testErrors()
+{
+    ExclusiveZone zone(m_syncObject);
+    return static_cast<int>(m_errors.size());
+}
 
 // Gets the number of detected failures.
-inline int TestResult::testFailures ()
-{ ExclusiveZone zone (m_syncObject); return m_failures.size (); }
-
+inline int TestResult::testFailures()
+{
+    ExclusiveZone zone(m_syncObject);
+    return static_cast<int>(m_failures.size());
+}
 
 // Returns whether the entire test was successful or not.
-inline bool TestResult::wasSuccessful ()
-{ ExclusiveZone zone (m_syncObject); return m_failures.size () == 0 && m_errors.size () == 0; }
-
+inline bool TestResult::wasSuccessful()
+{
+    ExclusiveZone zone(m_syncObject);
+    return m_failures.size() == 0 && m_errors.size() == 0;
+}
 
 // Returns a vector of the errors.
-inline std::vector<TestFailure *>& TestResult::errors ()
-{ ExclusiveZone zone (m_syncObject); return m_errors; }
-
+inline std::vector<TestFailure*>& TestResult::errors()
+{
+    ExclusiveZone zone(m_syncObject);
+    return m_errors;
+}
 
 // Returns a vector of the failures.
-inline std::vector<TestFailure *>& TestResult::failures ()
-{ ExclusiveZone zone (m_syncObject); return m_failures; }
-
+inline std::vector<TestFailure*>& TestResult::failures()
+{
+    ExclusiveZone zone(m_syncObject);
+    return m_failures;
+}
 
 // Returns whether testing should be stopped
-inline bool TestResult::shouldStop ()
-{ ExclusiveZone zone (m_syncObject); return m_stop; }
-
+inline bool TestResult::shouldStop()
+{
+    ExclusiveZone zone(m_syncObject);
+    return m_stop;
+}
 
 // Stop testing
-inline void TestResult::stop ()
-{ ExclusiveZone zone (m_syncObject); m_stop = true; }
-
+inline void TestResult::stop()
+{
+    ExclusiveZone zone(m_syncObject);
+    m_stop = true;
+}
 
 // Accept a new synchronization object for protection of this instance
 // TestResult assumes ownership of the object
-inline void TestResult::setSynchronizationObject (SynchronizationObject *syncObject)
-{ delete m_syncObject; m_syncObject = syncObject; }
-
+inline void TestResult::setSynchronizationObject(SynchronizationObject* syncObject)
+{
+    delete m_syncObject;
+    m_syncObject = syncObject;
+}
 
 #endif
-
-
